0.5.0

+++ install/install.sh
@@ -414,7 +414,8 @@ sleep 25
 echo "edit nginxproxymanager password..." 
 login_data=$(curl -X POST -H "Content-Type: application/json" -d '{"identity":"admin@example.com","scope":"user", "secret":"changeme"}' http://127.0.0.1:9092/api/tokens)
 sleep 3
-token=$(echo $login_data | jq -r '.token')
+token=echo $login_data | grep -Po '(?<="token":")[^"]*'
+#token=$(echo $login_data | jq -r '.token')
 new_password=$(docker run --name pwgen backplane/pwgen 15)!
 docker rm -f pwgen
 curl -X PUT -H "Content-Type: application/json" -H "Authorization: Bearer $token" -d '{"email": "help@websoft9.com", "nickname": "admin", "is_disabled": false, "roles": ["admin"]}'  http://127.0.0.1:9092/api/users/1
@@ -432,6 +433,8 @@ public_ip=`bash /data/apps/stackhub/scripts/get_ip.sh`
 sudo sed -i "s/domain.com/$public_ip/g" /var/lib/docker/volumes/w9nginxproxymanager_nginx_data/_data/nginx/proxy_host/initproxy.conf
 
 echo "Add htpasswd for appmanage ..."
+new_password=$(docker run --name pwgen backplane/pwgen 15)!
+docker rm -f pwgen
 sudo sed -i 's/"APPMANAGE_USERNAME": ".*"/"APPMANAGE_USERNAME": "websoft9"/g' /usr/share/cockpit/myapps/config.json
 sudo sed -i 's/"APPMANAGE_PASSWORD": ".*"/"APPMANAGE_PASSWORD": "'$new_password'"/g' /usr/share/cockpit/myapps/config.json
 rm -rf /var/lib/docker/volumes/w9nginxproxymanager_nginx_data/_data/nginx/proxy_host/.htpasswd /tmp/.htpasswd
@@ -444,28 +447,30 @@ sudo docker restart websoft9-nginxproxymanager
 EditMenu(){
 
 echo "Start to  Edit Cockpit Menu ..."
-if [ -e /usr/share/cockpit/systemd ]; then
-  jq  '. | .tools as $menu | .menu as $tools | .tools=$tools | .menu=$menu | del(.tools.services) | del(.menu.preload.services) | .menu.index = .tools.index | del(.tools.index) | .menu.index.order = -2' /usr/share/cockpit/systemd/manifest.json > /usr/share/cockpit/systemd/manifest.json.tmp
-  rm -rf /usr/share/cockpit/systemd/manifest.json
-  mv /usr/share/cockpit/systemd/manifest.json.tmp /usr/share/cockpit/systemd/manifest.json
-  cd /usr/share/cockpit/systemd && rm -rf services.js.gz services.html.gz services.css.gz
-fi
-if [ -e /usr/share/cockpit/networkmanager ]; then
-  sudo sed -i 's/menu/tools/g' /usr/share/cockpit/networkmanager/manifest.json
-fi
-if [ -e /usr/share/cockpit/storaged ]; then
-  sudo sed -i 's/menu/tools/g' /usr/share/cockpit/storaged/manifest.json
-fi
-if [ -e /usr/share/cockpit/users ]; then
-  sudo sed -i 's/menu/tools/g' /usr/share/cockpit/users/manifest.json
-fi
+if command -v yq > /dev/null; then
+  if [ -e /usr/share/cockpit/systemd ]; then
+    jq  '. | .tools as $menu | .menu as $tools | .tools=$tools | .menu=$menu | del(.tools.services) | del(.menu.preload.services) | .menu.index = .tools.index | del(.tools.index) | .menu.index.order = -2' /usr/share/cockpit/systemd/manifest.json > /usr/share/cockpit/systemd/manifest.json.tmp
+    rm -rf /usr/share/cockpit/systemd/manifest.json
+    mv /usr/share/cockpit/systemd/manifest.json.tmp /usr/share/cockpit/systemd/manifest.json
+    cd /usr/share/cockpit/systemd && rm -rf services.js.gz services.html.gz services.css.gz
+  fi
+  if [ -e /usr/share/cockpit/networkmanager ]; then
+    sudo sed -i 's/menu/tools/g' /usr/share/cockpit/networkmanager/manifest.json
+  fi
+  if [ -e /usr/share/cockpit/storaged ]; then
+    sudo sed -i 's/menu/tools/g' /usr/share/cockpit/storaged/manifest.json
+  fi
+  if [ -e /usr/share/cockpit/users ]; then
+    sudo sed -i 's/menu/tools/g' /usr/share/cockpit/users/manifest.json
+  fi
 
-jq  '. | del(.locales."ca-es") | del(.locales."nb-no") | del(.locales."sk-sk") | del(.locales."tr-tr")| del(.locales."cs-cz") | del(.locales."de-de") | del(.locales."es-es") | del(.locales."fi-fi") | del(.locales."fr-fr") | del(.locales."it-it") | del(.locales."ja-jp") | del(.locales."pl-pl") | del(.locales."pt-br") | del(.locales."ru-ru") | del(.locales."sv-se") | del(.locales."uk-ua") | del(.locales."zh-tw") | del(.locales."he-il") | del(.locales."nl-nl")  | del(.locales."ko-kr") | del(.locales."ka-ge")' /usr/share/cockpit/shell/manifest.json > /usr/share/cockpit/shell/manifest.json.tmp
-rm -rf /usr/share/cockpit/shell/manifest.json
-mv /usr/share/cockpit/shell/manifest.json.tmp /usr/share/cockpit/shell/manifest.json
-
+  jq  '. | del(.locales."ca-es") | del(.locales."nb-no") | del(.locales."sk-sk") | del(.locales."tr-tr")| del(.locales."cs-cz") | del(.locales."de-de") | del(.locales."es-es") | del(.locales."fi-fi") | del(.locales."fr-fr") | del(.locales."it-it") | del(.locales."ja-jp") | del(.locales."pl-pl") | del(.locales."pt-br") | del(.locales."ru-ru") | del(.locales."sv-se") | del(.locales."uk-ua") | del(.locales."zh-tw") | del(.locales."he-il") | del(.locales."nl-nl")  | del(.locales."ko-kr") | del(.locales."ka-ge")' /usr/share/cockpit/shell/manifest.json > /usr/share/cockpit/shell/manifest.json.tmp
+  rm -rf /usr/share/cockpit/shell/manifest.json
+  mv /usr/share/cockpit/shell/manifest.json.tmp /usr/share/cockpit/shell/manifest.json
+else
+  echo "system have no jq, use cockpit menu ..."
+fi
 echo "---------------------------------- Install success!  you can  install a app by websoft9's appstore -------------------------------------------------------" 
-
 }
 
 CheckEnvironment
